+++ otterdog/operations/sync_template_operation.py
@@ -14,7 +14,7 @@ from colorama import Style
 from otterdog.config import OrganizationConfig
 from otterdog.models.github_organization import GitHubOrganization
 from otterdog.providers.github import Github
-from otterdog.utils import is_set_and_valid
+from otterdog.utils import is_set_and_present
 
 from . import Operation
 
@@ -62,26 +62,21 @@ class SyncTemplateOperation(Operation):
                 if self._repo is not None:
                     if repo.name != self._repo:
                         continue
-                elif not is_set_and_valid(repo.template_repository):
+                elif repo.archived is True:
                     continue
 
-                if repo.archived is True:
-                    continue
+                if is_set_and_present(repo.template_repository):
+                    self.printer.print(f"Syncing {Style.BRIGHT}repository[\"{repo.name}\"]{Style.RESET_ALL}")
+                    updated_files = \
+                        gh_client.sync_from_template_repository(github_id,
+                                                                repo.name,
+                                                                repo.template_repository,
+                                                                repo.post_process_template_content)
 
-                ignore_paths = \
-                    repo.post_process_template_content if is_set_and_valid(repo.post_process_template_content) else []
-
-                self.printer.print(f"Syncing {Style.BRIGHT}repository[\"{repo.name}\"]{Style.RESET_ALL}")
-                updated_files = \
-                    gh_client.sync_from_template_repository(github_id,
-                                                            repo.name,
-                                                            repo.template_repository,
-                                                            ignore_paths)
-
-                self.printer.level_up()
-                for file in updated_files:
-                    self.printer.print(f"updated file '{file}'")
-                self.printer.level_down()
+                    self.printer.level_up()
+                    for file in updated_files:
+                        self.printer.print(f"updated file '{file}'")
+                    self.printer.level_down()
 
             return 0